Correctly materialise implicit relations (#5294)
## What is the goal of this PR?
Previously, when querying for generic relations, if we encounter a rule which inferred an attribute, we would convert it to an implicit relations. When materialising these, we could encounter situations when the insert would fail - an implicit relation where the attribute role player is only specified by the value and type. In this PR in such cases, for materialisation we convert it back to a resource so all information is at hand.

## What are the changes implemented in this PR?

- when materialising implicit relations, we materialise them as attributes if possible so that all information needed for insertion is at hand