Do not flush completeness when inserting inferred concepts (#5306)
## What is the goal of this PR?

Previously (#5300) we introduced cache flushing mechanism for insertions and deletions. The added behaviour flushes too often however - the completion is flushed on insertion of inferred concepts. Insertion of inferred concepts is an integral part of the reasoning and it doesn't require any extra mechanisms to ensure the query cache is up to date. As a result of flushing on insertion of inferred concepts we potentially need to do extra db checks to guarantee answer completion.

## What are the changes implemented in this PR?
- we do not ack a concept insertion in the query cache if the concept is inferred
- we reinstate the extra head satisfiability check