Remove the use of assert (#5332)
## What is the goal of this PR?

To remove the use of `assert` in the code. The motivation is that it is only meant to be used during development and it requires a VM option to be enabled during run time. Additionally, if the assert condition is false, the assert throws an AssertionError with just the message text.

## What are the changes implemented in this PR?
- Removes the use of `assert` and uses `Preconditions` in place.