Ensure rule states are created lazily (#5337)
## What is the goal of this PR?
Previously the rule state iterator was built from a stream containing a stateful intermediate operation - sorting, which requires processing of all elements. As a result the rule states were constructed eagerly. This PR makes sure that they are created lazily.

## What are the changes implemented in this PR?
Presorts the stream source and ensures no stateful ops are present in the stream of rule states.