Refactor and optimise resolution plan (#5338)
## What is the goal of this PR?
To simplify and correct the behaviour of resolution planning: remove the recursion and make sure the ordering ensures atom connectedness if possible.

## What are the changes implemented in this PR?
Context: when doing resolution we need possibly need to split queries into atomic ones so that we can resolve them via rules. The entry query is split based on the plan graql produces. The split queries are then executed separately.  However graql planner does not always produce ordering optimal for resolution purposes so we introduced a refinement step - it ensures we use the information in the query (ids, specific value predicates) optimally, we start from subqueries with the most amount of information. This PR addresses some issues of the refinement step.
- remove recursion in GraqlTraversalPlanner
- make sure GraqlTraversalPlanner doesn't produce unnecessarily disconnected plans by tracking visited variables
- produce and log a warning if created resolution plan is disconnected