include subtypes when determining fruitless rules (#5366)
## What is the goal of this PR?
Previously in the `RuleCache` we were trying to assess whether a type has instances. To do that we were looking whether the type has instances and whether an instance could be created via a rule. When looking at possible rules to do that, we were only looking at rules of the specific type omitting it's subtypes. This PR fixes that.

## What are the changes implemented in this PR?
Additionally check subtypes when trying to find matchable rules when determining whether an instance has types.