Look for non-trivial instance loops when determining whether reiteration is needed (#5373)
## What is the goal of this PR?
When trying to determine whether a resolution reiteration is needed, we check for cycles in the type dependency graph and later in the relation instances. Previously when looking for cycles among relations, we were looking at direct neighbours only. This would lead finishing the processing too early in some cases. Now we look for non-trivial (transitive) cycles.
## What are the changes implemented in this PR?
Uses Tarjan's algo to determine reachable neighbours when looking for cycles in relation instances. 
Fixes #5355