used persisted values when materialising attributes (#5380)
## What is the goal of this PR?
Previously when materialising attributes (inserting inferred attributes) we were using the value provided in the query. It was possible to have a mismatch between the provided value and the persisted value of the corresponding attribute. As a result, it was possible to miss an already existing attribute and proceed with the insertion resulting in an attribute duplicate. This PR fixes this issue by ensuring we are always following the PUT behaviour with a persisted attribute data type.

## What are the changes implemented in this PR?
When materialising attributes, before checking for attribute existence, we convert the value to the persisted data type