fix null pointer issue when retrieving meta implicit from ruel cache (#5403)
## What is the goal of this PR?
Querying for meta implicit relations: `$x isa @has-attribute`, results in a Null Pointer exception. 

The reason is that when trying to determine the rules that are applicable in this case we look at the downstream type hierarchy of `'has-attribute` including its explicit equivalents. However the meta implicit relation doesn't have an explicit equivalent and its db lookup results in a null.

## What are the changes implemented in this PR?
Added extra null stream filter when fetching types in `RuleCache`.