Use keyspace stats to fetch instance counts (#5420)
## What is the goal of this PR?
The goal of this PR is to optimise the `compute count` queries. 
Currently we are computing the counts each time the query is executed. This is potentially heavy and time-consuming for large graphs. However, we are already storing instance counts on type vertices in the graph. These come in the form of `KeyspaceStatistics`. We could readily use this information to provide instant compute count responses. 
In this PR we use the cached counts to provide instant response to compute count queries.

## What are the changes implemented in this PR?
Use `KeyspaceStatistics` counts in `runComputeCount` in `ComputeExecutor`.