Get rid of DisjunctionIterator (#5382)
## What is the goal of this PR?
We introduced the `DisjunctionIterator` to enforce sequentiality of processing of conjunctions to leverage query cache among conjunctions. 

What it does is it creates an iterator for each component conjunction of the provided disjunction, from which it provides a top level merging iterator for the whole query.

Since `QueryCache` has now a transaction lifetime, this extra handling is not needed. We can simply use sequential streams within the same transaction.
Additionally, the `DisjunctionIterator` was preventing us from problems with `flatMap` stream operation (https://bugs.openjdk.java.net/browse/JDK-8075939). Since we introduced Vavr to handle that, this use case is also redundant.

## What are the changes implemented in this PR?
Depends on #5383 
- Remove `DisjunctionIterator`
- use native reasoner methods and Vavr streams to provide the adequate logic in `QueryExecutor`