Disable release-validate-deps since we have a cyclic dependency where client-java depends on core, core depends on console and console depends on client-java (#5444)
## What is the goal of this PR?

We temporarily disable release-validate-deps for console in order to allow for releasing grakn without having to release console first. Otherwise, it would be impossible to release grakn, console, and client-java due to the cyclic dependency that exist between them. The issue for removing the cyclic dependency is tracked here: https://github.com/graknlabs/grakn/issues/5272.