Implement Custom Lazy Stream Merging (#5449)
## What is the goal of this PR?
Replace io.vavr (#5430) that was introduced to solve the lack of lazy `flatMap` that is mentioned in #5430. We then determined that conversion to the `io.vavr` Stream implementation is more lazy, but the conversion step itself does a single call to `.next()`. In practical terms, this means Grakn always precomputed the first answer on query, which is still not fully lazy.
This led to the decision to restore prior behavior and solve the eagerness problem by implementing our own lazy stream merging solution.

## What are the changes implemented in this PR?
* Implement a lazy stream merging solution that uses `stream.iterator()` conversions interally
* Remove `io.vavr` as a dependency
* tests for lazy stream merging