Retrievable Explanations (#5483)
## What is the goal of this PR?
Large explanations can break gRPC message limits, specifically nesting depth when returning deep explanation trees. With this change, each layer of the explanation tree can be retrieved from the server separately. This reflects the changes in the protocol here: https://github.com/graknlabs/protocol/pull/20. 

We retain the storage of Explanation trees on the Server, and expose a new RPC message with corresponding handler on the server to retrieve particular layers of the tree using the query pattern provided by the client. The query pattern is used to access the query cache directly.

Additionally, `pattern` in `ConceptMap`s now contain the IDs indicated in the actual concept map as well.

Finishes #4605, and #4545 

## What are the changes implemented in this PR?
* `Pattern` has been moved from `Explanation` to `ConceptMap`
* `Answer` interface is empty and requires further refactoring
* Propagate `Pattern` throughout reasoner conclusions
* Reconstruct `Explanation` objects from the query cache when the user requests an explanation a concept map
* Update some tests
* Solve previously unnoticed explanation bug: when re-querying the same query, the explanation is shallow (ie. a `LookupExplanation`) rather than the full reasoner tree