Replace __.identity() noop with __.start() even cheaper anonymous traversal (#5507)
## What is the goal of this PR?
A minor but free performance enhancement for `read` queries, in which we often have `__.identity()` steps in the traversal. These are redundant (documented no-op in janus) and were generated by our need to create anonymous traversals that could be joined to existing ones. Instead, we can use `__.start()` to obtain a new traversal. Since the `identity()` method actually iterates over all elements in the traversal (it appears, because it takes non-zero time) we want to avoid it.

## What are the changes implemented in this PR?
* Replace `__.identity()` noop with `__.start()`, an even cheaper anonymous traversal