Selective locking on commit (#5492)
## What is the goal of this PR?
When loading concurrently, we want to minimise the amount of locking needed to improve performance whilst retaining correctness and consistency.

Currently we lock the graph if there are any attribute changes in any of the committing transactions. This in practice means that we lock practically always and all the time for mixed loads. This is however very defensive and not required as the only situations that require locking on commit are when two transactions share an inserted attribute or want to create a shard vertex for the same type.

To be able to lock more effectively, we need to be able to a) lock selectively based on the creation of identical attributes among different transaction, b) lock selectively based on the creation of shards of the same type among different transaction. Hence we introduce a basic selective locking behaviour.

## What are the changes implemented in this PR?
- if two transactions insert the same attribute, we require them to acquire graph locks when committing. To be able to recognise this situation taking place, we introduce the `AttributeManager`. The `AttributeManager` is a session-wide object used to manage attribute mutations. All transactions must report their attribute insertions, deletions and commits to the `AttributeManager`. The `AttributeManager` then tracks the attribute mutations as well as txs in which the mutations took place. By logging this information the `AttributeManager` can then find and recognise possible contention. The contention resolution happens in the following way. When about to commit, each transaction polls the `AttributeManager` if it needs to use a lock during commit. If the `AttributeManager` finds at least two transactions with a shared attribute, it will advise the competing transactions to use a lock when committing.

- if two transactions are about to insert a shard for the same type, we require them to acquire graph locks when committing. To do so we introduce the `ShardManager`. The `ShardManager` is a session-wide object used to resolve shard creation contention. Just before committing, all transactions are required to signal their need to create a new shard vertex to the `ShardManager`. The `ShardManager` then tracks the transaction shard requirements for specific types. This way it can find and resolve possible contention. When a transaction is about to commit, it polls the `ShardManager` if it needs to use a lock during commit. If the `ShardManager` finds at least two transactions with a shared shard vertex request, it will advise the competing transaction to use a lock when committing.