Ensure query cache doesn't return duplicate answers (#5520)
## What is the goal of this PR?
We had a mismatch between what `SessionService` was expecting from the `QueryCache`, namely it didn't expect it to return duplicate answers. Duplicate answers could have been returned in cases when the entry query had some inferred entries in the cache but the entry wasn't dbComplete. In that case we would mix a full DB fetch of answers with what we have in cache. We correct the mismatch by making sure `QueryCache` doesn't return duplicates when mixing these.
## What are the changes implemented in this PR?
Fixes #5518
- added a `distinct` to the mixed stream when doing answer fetch in `QueryCache`
- fixed an issue with recognising whether query cache contains an answer to a ground query