Split GraknTestServer (#5532)
## What is the goal of this PR?
When performing integration tests for Grakn's components such as reasoner, queries... we should not be starting the full Grakn server and extracting parts of the running server. Instead, we should start Cassandra and build only the required components for the tests we wish to run. This change enables us to start Cassandra as a test rule separately from the full `GraknTestServer`.

## What are the changes implemented in this PR?
* Changes the integration testing paradigm, in that we can start Storage via a new components `GraknTestStorage` without starting `GraknTestServer`
* Enlarge a util class `SessionUtil` to produce sessions for tests, allowing us to inject and customise each test as required
* Clean up imports, add some comments

Finishes #5529