Revise query cache behaviours (#5537)
## What is the goal of this PR?

To revise the query cache behaviours so that we can increase the cache efficiency. In certain scenarios query results weren't reused when they could have. The specific scenario we address here is when there is a subsumption relation between a cache query and input query up to the choice of ids - which we refer to as structural subsumption.

Previously, we stored query cache entries with specific IDs attached in the lookup key. This led to a second layer of stored IDs that are all valid substitutions into the query. Effectively, the specific IDs attached at the lookup key are placeholders that we now ignore (using structural subsumption), and so can find all the IDs matching a query rather than a specific one.

## What are the changes implement in this PR?
* introduced `STRUCTURAL_SUBSUMTIVE` unifier for determination of structural subsumption relation
* query cache entries are now compared wrt structural subsumption
* revised the `GET` and `PROPAGATE` behaviours to incorporate the new comparison