Fix rule overpruning when rules with negated patterns do not have instances (#5558)
## What is the goal of this PR?
When processing rules in `RuleCache`, a rule is discarded if any of its body types do not have an instance in the graph. This is because in such a scenario the rule cannot yield any answers. The instances with no instances are tracked in the `RuleCache`. The current approach takes into account that a premise might be negated - we only consider positive premises. However! The approach doesn't take into account the scenario in which 2 rules with the same type are present where one is positive and one is negative. In such scenario, if the type doesn't have instances and the rule with the positive type gets processed first, when processing the rule with negative type we would discard it.

## What are the changes implemented in this PR?
- when considering types with no instances, we make sure that any possibly discarded rule doesn't negate the premise with that type
- went from Multimap to Map so that we can differentiate whether a type has not been visited (get returns null) or has been visited but no rules are applicable (get returns an empty set)
- moved the whole logic of rule acking to `RuleCache`