Also log the type of a Thing that needs a key (#5583)
## What is the goal of this PR?

We have improved the error message that is displayed when you're not inserting the key of a concept that needs it. Not knowing the type of the concept you are trying to insert as you try to fix the query is quite painful.

The error message before this PR:
```
Thing [V4320] does not have exactly one key of type [symbol]. It either has no keys assigned to it, or it has more than one. 
```

The error message **after** this PR:
```
Thing [V4320] of type [commit] does not have exactly one key of type [symbol]. It either has no keys assigned to it, or it has more than one. 
```