Introduce reasoning context (#5597)
## What is the goal of this PR?
To make atom construction and related classes cleaner. Before we were passing a number of context arguments to bottom classes which were propagated among the hierarchy. Now we abstract the context arguments into a single class which can then be stored in the base `Atom` class.

## What are the changes implemented in this PR?
We introduce a container `ReasoningContext` class which stores a `ReasonerQueryFactory`, `ConceptManager`, `RuleCache`, `QueryCache`, `KeyspaceStatistics`.