Store labels instead of ids in atoms (#5612)
## What is the goal of this PR?
To store `SchemaConcept` labels instead of their ids in atoms. 
Reasons: 
1) ids carry little information and require further manipulation (lookups) to extract their meaning
2) we often want to use the label which then requires doing a lookup via `ConceptManager`. 

As we trim down the number of dependencies on `ConceptManager`, this change also should allow to remove `ReasoningContext` out of atoms and instead inject it to functions that require it.

## What are the changes implemented in this PR?
Instead of storing type `ConceptId`s in atoms, we store their `Label`s.