Flatten RelationAtom (#5620)
## What is the goal of this PR?
Instead of extending `IsaAtom`, we make the `RelationAtom` contain the associated `IsaAtom`. As a result the Atom hierarchy is more flat and simple and we enable further simplifications.

## What are the changes implemented in this PR?
`RelationAtom` now simply extends `Atom`. We store the binary information in a appropriate `IsaAtom` which we use directly for relation unification and equivalence.