Extract TraversalExecutor from QueryExecutor (#5640)
## What is the goal of this PR?

Refactor `QueryExecutorImpl`, which is instantiated per-query with an `infer` flag. Interally it used to perform traversals on the backend, which was completely independent of inference. This PR extracts `TraversalExecutor` out of `QueryExecutor`, which has the side effect of removing `ExecutorFactory` from much of reasoner and only passing a `TraversalExecutor` directly. This makes more sense than before, as the semantics of holding a `ExecutorFactory` indicates reasoner might perform read, write, or compute queries where it really only needs to perform read traversals using the `TraversalFactory`.

## What are the changes implemented in this PR?
* Split `QueryExecutor` into `QueryExecutor` and `TraversalExecutor`
* `TraversalExecutor` is a dedicated, lower level component that only consumes patterns and executes traversals on the backend without any knowledge of reasoning