Defer db completeness check in query cache when fetching (#5673)
## What is the goal of this PR?
When fetching answers from the query cache, we are currently checking for query completeness at the very beginning of the call. What is possible however is that we propagate answers from a complete parent and hence acknowledge a child complete. As a result, currently we would still resort to doing a DB lookup where all answers are actually available in the cache. Here, we correct this behaviour by deferring the completeness check to after potential answer propagation.

## What are the changes implemented in this PR?
Check for db-completeness only after we analysed the query match in `getAnswerStreamWithUnifier` in `SemanticCache`.