Fix issue with too eager subgoal propagation for queries with negation (#5699)
## What is the goal of this PR?
Negation queries are short circuited when one contradiction is found from the negation block. However, we mark negation sub-queries as fully explored by adding the sub-query to the visited `subgoals` that is shared globally. We then re-visit the same state at another point while exploring the negation query. However, we are missing answers, since the cache will only contain a single answer to the visited sub-query, not all of them. This means we can miss other solutions to the negation block. This is the underlying problem of #5697.

## Detailed Explanation


The issue can be described in terms of the 2 following query profiles which both describe the execution of a variant of the query:

```
($flt, $ques) isa fault-identification;
(question-not-asked: $ques, parent-session: $ts) isa unasked-question;
$flt id ...;
$ts id ...;
```
inside a negation block, for different substitutions for `$flt`.

The first profile shows the first execution of the query for `[$flt/65736]`. The conjunctive query is decomposed into atomic queries, the first query is:

`unasked-question($ts, $ques) [$ts/V53256]`

The query is seen for the first time (not been visited before) hence a rule is applied. Since the query is within a negation block, once an answer is found, we immediately return.

![image](https://user-images.githubusercontent.com/10519067/80311878-1ef1ba00-87da-11ea-9bff-f64e34dbeb19.png)

The second query shows the execution of a following query with  `[$flt/32960]`. Once we decompose the conjunction, the query produces the same query as in the first profile:

`unasked-question($ts, $ques) [$ts/V53256]`
This query has however been marked as visited and we haven't consumed all its answers as we were interested only in a single one. A quick inspection of the db reveals that there is actually a pair. As a result we only do a table lookup and retrieve a single answer which is then fed into the second atomic query. The atomic query correctly doesn't yield any answers. We have however not processed the second answer from the pair which is the one that satisfies the conjunction. An answer is lost.
![image](https://user-images.githubusercontent.com/10519067/80311889-2749f500-87da-11ea-852d-70bba48076ef.png)

## What are the changes implemented in this PR?

Make sure we do not reuse global subGoals in negation branches - we do an explicit copy before branching out.

Fixes #5697