Block gRPC request observer thread to get back-pressure (#5728)
## What is the goal of this PR?

Fix #5726 by allowing gRPC's natural back-pressure (automatic flow control) to take place.

## What are the changes implemented in this PR?

Request messages are no longer eagerly queued asynchronously for processing in the TransactionService, and instead are blocked by waiting on the returned future when submitting them. This blocks the gRPC thread, which prevents the server buffer from being depleted and allows gRPC to control the back-pressure.

Also fixed dependency on a now non-existent commit of verification (due to a rebase).