Change behaviour of explanations and their patterns (#5723)
## What is the goal of this PR?

We change the behaviour of explanations quite drastically. Previously explanations were only provided to explain answers derived using a rule. Now we introduce explanations for negations and disjunctions, which means that an explanation will be present for any query that contains a disjunction or negation.

Explaining a disjunction now shows the clause of the disjunction that yielded the answer, including the full ConceptMap for that clause.

We alter the format of patterns in answers. Patterns still include ids, which should be limited to the set of ids that are returned in the answer's ConceptMap. These ids now only occur once each within the pattern, and will always sit in the root conjunction of the pattern.

Previously patterns for conjunctive queries were working, but patterns for queries and rules with negation were incorrect. This is necessary for our efforts to build a resolution testing framework.

## What are the changes implemented in this PR?

- Ids in answers' patterns now occur at most once, and in the outermost scope of the pattern.
- Changes the behaviour of how negations and disjunctions are represented in explanations
  - Disjunctions are now represented in the pattern in their DNF'ed form, and explained with the clause that was matched for the given answer
  - Negations now include the omitted `not`, and are explained by the lookup that was made (prior to being checked against the `not` clause)