catch all exception (#5820)
## What is the goal of this PR?

Currently, not all exceptions thrown in tracing handling are caught and close the connection properly. Some exceptions are thrown but swallowed, which means we don't see any error reporting when an exception happens.

## What are the changes implemented in this PR?

Raise the closing connection logic to a higher level so that whenever an exception is thrown, we always close the connection.