Add role downward inference when getting relations or castings (#5864)
## What is the goal of this PR?
To ensure that the methods in Concept API present a consistent behaviour, we implement downward type inference when retrieving role players or relations over a given role. We always perform downward type inference in Concept API read operations.

Before, we only searched for role players or relations playing the exact role without checking the role subtypes. This aligns with attribute ownership and owned-ness - we can ask for `name` owned by a thing and we get back `first-name` and `last-name` instances as well. 

## What are the changes implemented in this PR?
* Implement a call to `role.subs()` when retrieving role players
* Implement a call to `role.subs()` when retrieving relations that a thing plays a role in
* Correspondingly update the role player delete executors