Fix TransactionRPC.getConcept returning deleted concepts (#10)
## What is the goal of this PR?

To fix https://github.com/graknlabs/grakn-2.0/issues/9: "After deleting a concept, it is still returned when we look it up by IID".

## What are the changes implemented in this PR?

We created a new method `Graph.get(VERTEX_IID)` which behaves similarly to `Graph.convert`, but instead of assuming that the given IID exists in the graph, it instead first tries to look up the IID from our cache, then attempts retrieval from Storage; and if it is in neither, then it returns `null`.

However, simply changing `Concepts.getConcept` to use `get` instead of `convert` caused a different error; `ConceptRPC.convert` does in fact assume that conversion is performed, even if the IID passed in does not already exist in either Storage or our cache. So we introduced `Concepts.convertConcept` to ensure that this method still functions as intended.